requirements.txt
flask==3.0.0
flask-pymongo==2.3.0
pymongo==4.6.1
werkzeug==3.0.1
pyjwt==2.8.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
